requirement.txt
audioread==2.1.8
bs4==0.0.1
gTTS==2.1.0
gTTS-token==1.1.3
Pillow==6.2.0
python-vlc==3.0.7110
SpeechRecognition==3.8.1
